git-svn-id: http://managed-x86.googlecode.com/svn/trunk@10 14dbf6ce-8ea5-11dd-98a5-cbe9a756ba3b

+++ ControlFlow/X86Writer.Patch.cs
@@ -21,9 +21,9 @@ namespace Managed.X86 {
 		/// <param name="target">The target address</param>
 
 		public void Patch(IntPtr target) {
-			uint disp, size = 0;
+			int disp, size = 0;
 			int peek = stream.ReadByte();
-			uint pos = (uint)this.Position.ToInt32();
+			int pos = this.Position.ToInt32();
 			switch (peek) {
 				#region call, jump32
 				case 0xe8:
@@ -49,7 +49,7 @@ namespace Managed.X86 {
 						throw new NotSupportedException();
 					}
 			}
-			disp = ((uint)target.ToInt32()) - pos;
+			disp = (target.ToInt32()) - pos;
 			if (size != 0) imm_emit32(disp - 4);
 			else if (is_imm8(disp - 1)) imm_emit8(disp - 1);
 			else throw new NotSupportedException();
